+++ crates/bcr-wallet-cli/src/command.rs
@@ -3,8 +3,6 @@ use bcr_wallet_core::AppState;
 use chrono::{DateTime, Utc};
 use tracing::info;
 
-use crate::WalletSettings;
-
 pub async fn cmd_info(app_state: &AppState) -> Result<String> {
     let mut res = String::new();
     let wallet_ids = app_state.get_wallets_ids().await?;
@@ -60,38 +58,27 @@ pub async fn cmd_info(app_state: &AppState) -> Result<String> {
     Ok(res)
 }
 
-pub async fn cmd_add_wallet(
-    app_state: &AppState,
-    settings: &WalletSettings,
-    name: &str,
-) -> Result<String> {
+pub async fn cmd_add_wallet(app_state: &AppState, name: &str) -> Result<String> {
     let mut res = String::new();
-    let id = app_state
-        .add_wallet(
-            name.to_owned(),
-            settings.mint_url.to_owned(),
-            settings.mnemonic.to_owned(),
-        )
-        .await?;
+    let id = app_state.add_wallet(name.to_owned()).await?;
     push_break(&mut res);
     push_break(&mut res);
     res.push_str(&format!("Created Wallet for {name} - Wallet ID: {id}.\n"));
     Ok(res)
 }
 
-pub async fn cmd_restore_wallet(
-    app_state: &AppState,
-    settings: &WalletSettings,
-    name: &str,
-) -> Result<String> {
+pub async fn cmd_delete_wallet(app_state: &AppState, name: &str, id: usize) -> Result<String> {
     let mut res = String::new();
-    let id = app_state
-        .restore_wallet(
-            name.to_owned(),
-            settings.mint_url.to_owned(),
-            settings.mnemonic.to_owned(),
-        )
-        .await?;
+    app_state.delete_wallet(id).await?;
+    push_break(&mut res);
+    push_break(&mut res);
+    res.push_str(&format!("Deleted Wallet for {name} - Wallet ID: {id}.\n"));
+    Ok(res)
+}
+
+pub async fn cmd_restore_wallet(app_state: &AppState, name: &str) -> Result<String> {
+    let mut res = String::new();
+    let id = app_state.restore_wallet(name.to_owned()).await?;
     push_break(&mut res);
     push_break(&mut res);
     res.push_str(&format!("Restored Wallet for {name} - Wallet ID: {id}.\n"));